requirements.txt
# Use CPU-only or Apple Silicon compatible PyTorch versions
torch==2.0.1
torchvision==0.15.2
mlflow>=2.0.0,<3.0.0
optuna>=3.0.0
numpy>=1.21.0,<2.0.0
pillow>=8.3.0
scikit-learn>=1.1.0,<1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.64.0
kagglehub>=0.2.0
# Version constraints for Apple Silicon compatibility
pydantic>=1.8.0,<2.0.0
flask>=2.0.0,<3.1.0
jinja2>=3.0.0,<3.2.0